requirements.txt
psutil==5.2.2
aiohttp==2.2.0
attrs==17.2.0
structlog[dev]==17.2.0
pykube==0.15.0
pytest==3.1.2
